Revert "try moving LinCircle"
This reverts commit a9d4d8efb1434bb82bd45f3ee310b91d87f5a42e.
